refactor: generalize tenant service implementation
Move the IdentityTenantService class to the gateway module
so that it can be re-used by both the gRPC and REST API
modules.
